requirements.txt
Django==5.1.1
djangorestframework==3.15.2
django-storages==1.14.4
uuid~=1.30
drf-yasg~=1.21.7
python-dateutil~=2.8.2
python-dotenv~=1.0.1
requests~=2.31.0